fix: resolve compiler warnings and CodeQL constant-comparison
- IccHeuristicsDataValidation.cpp: fix sign-compare warning (int vs uint32)
- IccHeuristicsProfileCompliance.cpp: mark unused profilePath parameter
- IccHeuristicsXmlSafety.cpp: mark unused pIcc parameter
- IccHeuristicsRawPost.cpp: use ifSelCount for CWE-400 branching opcode detection
- IccImageAnalyzer.cpp: remove redundant tileW>0 check (CodeQL constant-comparison)

Verified:
[OK] build.sh: 0 warnings (was 4)
[OK] scan-build-18: 0 bugs (colorbleed + analyzer)
[OK] CodeQL: 11 baseline alerts, 0 new (203 queries)
[OK] Unit tests: 229/230 (1 pre-existing upstream HUAF)
[OK] MCP tests: 1811/1811 in 41s
[OK] WebUI tests: 256/256 in 17s
[OK] Docker build: exit 0, /api/health → {tools:24}
[OK] Docker /api/registry: 150 heuristics, 87 CVEs, 95 GHSAs
[OK] Docker upload+analyze: sRGB_D65_MAT.icc → 145 run, 136 OK
[OK] WebUI HTML: validate_xml lists 43 XML files, 5 tool forms verified

+++ iccanalyzer-lite/IccHeuristicsDataValidation.cpp
@@ -463,7 +463,7 @@ printf("[H66] Comprehensive NumArray NaN/Inf Scan\n");
                ColorCritical(), ColorReset());
         nanIssues++;
       }
-      if (extremeCount > scanLimit / 4) {
+      if (extremeCount > static_cast<int>(scanLimit / 4)) {
         printf("      %s[WARN]  Tag '%s': %d/%u extreme values (>1e10)%s\n",
                ColorWarning(), info.GetTagSigName(sit->TagInfo.sig),
                extremeCount, scanLimit, ColorReset());
